# mobile/docs/apis/openapi.json
{
    "openapi": "3.1.0",
    "info": {
      "title": "Fridge2Fork API",
      "description": "냉장고 재료 기반 한식 레시피 추천 API",
      "version": "1.0.0"
    },
    "servers": [
      {
        "url": "/",
        "description": "Default server"
      },
      {
        "url": "/fridge2fork",
        "description": "Kubernetes ingress path"
      }
    ],
    "paths": {
...
      "/fridge2fork/v1/fridge/ingredients": {
        "get": {
          "tags": [
            "냉장고"
          ],
          "summary": "Get Ingredients",
          "description": "전체 재료 목록 조회",
          "operationId": "get_ingredients_fridge2fork_v1_fridge_ingredients_get",
          "parameters": [
            {
              "name": "search",
              "in": "query",
              "required": false,
              "schema": {
                "anyOf": [
                  {
                    "type": "string"
                  },
                  {
                    "type": "null"
                  }
                ],
                "description": "재료명 검색",
                "title": "Search"
              },
              "description": "재료명 검색"
            },
            {
              "name": "page",
              "in": "query",
              "required": false,
              "schema": {
                "type": "integer",
                "minimum": 1,
                "description": "페이지 번호",
                "default": 1,
                "title": "Page"
              },
              "description": "페이지 번호"
            },
            {
              "name": "size",
              "in": "query",
              "required": false,
              "schema": {
                "type": "integer",
                "maximum": 200,
                "minimum": 1,
                "description": "페이지 크기",
                "default": 50,
                "title": "Size"
              },
              "description": "페이지 크기"
            }
          ],
          "responses": {
            "200": {
              "description": "Successful Response",
              "content": {
                "application/json": {
                  "schema": {
                    "type": "object",
                    "title": "Response Get Ingredients Fridge2Fork V1 Fridge Ingredients Get"
                  }
                }
              }
            },
            "422": {
              "description": "Validation Error",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/HTTPValidationError"
                  }
                }
              }
            }
          }
        },
        "post": {
          "tags": [
            "냉장고"
          ],
          "summary": "Add Ingredients",
          "description": "냉장고에 재료 추가 (세션 기반)",
          "operationId": "add_ingredients_fridge2fork_v1_fridge_ingredients_post",
          "requestBody": {
            "required": true,
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AddIngredientsRequest"
                }
              }
            }
          },
          "responses": {
            "200": {
              "description": "Successful Response",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/AddIngredientsResponse"
                  }
                }
              }
            },
            "422": {
              "description": "Validation Error",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/HTTPValidationError"
                  }
                }
              }
            }
          }
        },
        "delete": {
          "tags": [
            "냉장고"
          ],
          "summary": "Remove Ingredients",
          "description": "냉장고에서 재료 제거",
          "operationId": "remove_ingredients_fridge2fork_v1_fridge_ingredients_delete",
          "requestBody": {
            "required": true,
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RemoveIngredientsRequest"
                }
              }
            }
          },
          "responses": {
            "200": {
              "description": "Successful Response",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/RemoveIngredientsResponse"
                  }
                }
              }
            },
            "422": {
              "description": "Validation Error",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/HTTPValidationError"
                  }
                }
              }
            }
          }
        }
      },
...
    },
    "components": {
      "schemas": {
        "AddIngredientsRequest": {
          "properties": {
            "session_id": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Session Id"
            },
            "ingredients": {
              "items": {
                "additionalProperties": {
                  "type": "string"
                },
                "type": "object"
              },
              "type": "array",
              "title": "Ingredients"
            }
          },
          "type": "object",
          "required": [
            "ingredients"
          ],
          "title": "AddIngredientsRequest",
          "description": "재료 추가 요청"
        },
        "AddIngredientsResponse": {
          "properties": {
            "message": {
              "type": "string",
              "title": "Message"
            },
            "added_count": {
              "type": "integer",
              "title": "Added Count"
            },
            "session_id": {
              "type": "string",
              "title": "Session Id"
            }
          },
          "type": "object",
          "required": [
            "message",
            "added_count",
            "session_id"
          ],
          "title": "AddIngredientsResponse",
          "description": "재료 추가 응답"
        },
...
        "HTTPValidationError": {
          "properties": {
            "detail": {
              "items": {
                "$ref": "#/components/schemas/ValidationError"
              },
              "type": "array",
              "title": "Detail"
            }
          },
          "type": "object",
          "title": "HTTPValidationError"
        },
        "RemoveIngredientsRequest": {
          "properties": {
            "session_id": {
              "type": "string",
              "title": "Session Id"
            },
            "ingredients": {
              "anyOf": [
                {
                  "items": {
                    "type": "string"
                  },
                  "type": "array"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Ingredients"
            }
          },
          "type": "object",
          "required": [
            "session_id"
          ],
          "title": "RemoveIngredientsRequest",
          "description": "재료 제거 요청"
        },
        "RemoveIngredientsResponse": {
          "properties": {
            "message": {
              "type": "string",
              "title": "Message"
            },
            "removed_count": {
              "type": "integer",
              "title": "Removed Count"
            }
          },
          "type": "object",
          "required": [
            "message",
            "removed_count"
          ],
          "title": "RemoveIngredientsResponse",
          "description": "재료 제거 응답"
        },
        "ValidationError": {
          "properties": {
            "loc": {
              "items": {
                "anyOf": [
                  {
                    "type": "string"
                  },
                  {
                    "type": "integer"
                  }
                ]
              },
              "type": "array",
              "title": "Location"
            },
            "msg": {
              "type": "string",
              "title": "Message"
            },
            "type": {
              "type": "string",
              "title": "Error Type"
            }
          },
          "type": "object",
          "required": [
            "loc",
            "msg",
            "type"
          ],
          "title": "ValidationError"
        }
      }
    }
  }